The big rename (Transformers). Fixes #23.
